target/i386/sev: Add support for an SVSM
Add a new machine property to provide SVSM support. The new svsm option
points to a file that contains the SVSM binary.

The SVSM binary is initialized before the system firmware. The SVSM binary
lives in a memslot with a GPA outside of the guest memory limit, as
defined by the SVSM GPA and SVSM memory size found in the SVSM binary
GUIDed table. After loading the SVSM, a LAUNCH_UPDATE is performed to
encrypt and measure the SVSM at the VMPL0 level. The CPUID page and the
Secrets page LAUNCH_UPDATE calls are applied to the SVSM binary locations.

The guest system firmware continues to be encrypted and measured, but with
some changes if an SVSM is being used:
  - The system firmware is encrypted and measured at the VMPL1 level
  - The BSP register state is no longer obtained from the system firmware
    GUIDed table
  - The CPUID page and Secrets page locations are now treated as type zero
    pages during LAUNCH_UPDATE

The system firmware GPA is now provided in the Qemu firmware configuration
interface so that the SVSM can locate and parse the firmware information
more easily in order to locate the firmware GUIDed table information.

With the new BSP register state being supplied by the SVSM binary, the
guest will begin execution in the SVSM instead of the system firmware.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>